removed monitors.py
